backend/requirements.txt
idna==2.8
pylru>=1.0.6
tornado==4.4.3
python-dateutil==2.4.2
pytz
tzlocal
pysha3==1.0.2
simplejson>=3.6.0
xlrd
jinja2
openpyxl
pyoo
xlsxwriter
pillow
qrcode
rarfile
lxml
bandit==1.7.1
pytils==0.3
dbf==0.99.1
reportlab
natsort
protobuf==3.19.0
riemann_client==6.5.0
requests
mypy-lang
beautifulsoup4==4.7.1

raven==6.10.0
pyyaml==5.3.1
chardet==3.0.4

pyst2==0.4.9

kombu==4.2.0
celery==4.1.1
flower==0.9.2
redis==2.10.5

rest_condition
markdown
Django==2.2
django-rest-framework-mongoengine==3.4.1
mongoengine==0.18
pymongo==3.7.2
blinker==1.4
djangorestframework==3.11.0
django-filter==1.1.0

uwsgi
hiredis
django-redis-cache==1.7.1
django-extensions


zeep==3.4.0
srp==1.0.14
asn1crypto==1.3.0
uwsgitop

dill==0.3.0
jsonpickle==1.2
tenacity==6.0.0
transliterate

# Testing
pytest==5.4.3
pytest-html==2.1.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3
pytest-cov==2.10.0
pytest-xdist==1.32.0


# Linting
wemake-python-styleguide==0.14.1
flake8==3.9.2
pycodestyle==2.7.0

# Type checking
mypy==0.790
django-stubs==1.7.0
djangorestframework-stubs==1.3.0

# Logging & APM
django-structlog==1.6.1
python-logstash==0.4.6
elastic-apm==5.8.1

# Dev
coverage==5.3.1
django-split-settings==1.0.1
python-decouple==3.3
drf_yasg==1.15

django-prometheus==2.0.0
prometheus-client==0.7.1
